relocate secret pkg
